refactor Installers
